requirements.txt
robotframework
robotframework-seleniumlibrary
schedule
yagmail
Pillow
pyautogui
load_dotenv